refactor: 이름 분리 책임을 뷰로 이동
