+++ src/main/java/racingcar/view/CarNameSeparator.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.view;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ public class CarNameSeparator {
 
     private static final String SEPARATOR = ",";
 
-    public static List<String> separateNames(String carNames) {
+    public static List<String> separate(String carNames) {
         return List.of(carNames.split(SEPARATOR));
     }
