+++ src/test/java/racingcar/model/CarNameSeparatorTest.java
@@ -1,6 +1,7 @@
 package racingcar.model;
 
 import org.junit.jupiter.api.Test;
+import racingcar.view.CarNameSeparator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,7 +9,7 @@ public class CarNameSeparatorTest {
 
     @Test
     public void separateNamesComma() {
-        assertThat(CarNameSeparator.separateNames("apple,boost,banana"))
+        assertThat(CarNameSeparator.separate("apple,boost,banana"))
                 .containsExactly("apple", "boost", "banana");
     }
 }
